+++ bone-client/app/api/auth/[...nextauth]/route.ts
@@ -1,4 +1,7 @@
-import { authOptions } from "@/lib/auth";
-import NextAuth from "next-auth";
-const handler = NextAuth(authOptions);
-export { handler as GET, handler as POST };
+// import { authOptions } from "@/lib/auth";
+// import NextAuth from "next-auth";
+// const handler = NextAuth(authOptions);
+// export { handler as GET, handler as POST };
+
+import { handlers } from "@/lib/auth";
+export const { GET, POST } = handlers;
