+++ bone-client/app/page.tsx
@@ -2,54 +2,14 @@
 
 import React from "react";
 import Home from "../components/Home";
-import { signOut, useSession } from "next-auth/react";
-import Link from "next/link";
-import { useRouter } from "next/navigation";
 
 export default function HomePage() {
-  const { status } = useSession();
-  const router = useRouter();
-
-  const showSession = () => {
-    if (status === "authenticated") {
-      return (
-        <button
-          className="border border-solid border-black rounded"
-          onClick={() => {
-            signOut({ redirect: false }).then(() => {
-              router.push("/");
-            });
-          }}
-        >
-          Sign Out
-        </button>
-      );
-    } else if (status === "loading") {
-      return <span className="text-[#888] text-sm mt-7">Loading...</span>;
-    } else {
-      return (
-        <Link
-          href="/login"
-          className="border border-solid border-black rounded"
-        >
-          Sign In
-        </Link>
-      );
-    }
-  };
   return (
-    <main className="flex min-h-screen flex-col items-center justify-center">
-      <h1 className="text-xl">Home</h1>
-      {showSession()}
-    </main>
+    <>
+      {/* className="text-9xl text-blue-900" */}
+      <div>
+        <Home />
+      </div>
+    </>
   );
 }
-
-// return (
-//   <>
-//     {/* className="text-9xl text-blue-900" */}
-//     <div>
-//       <Home />
-//     </div>
-//   </>
-// );
