+++ bone-client/app/provider.tsx
@@ -1,6 +1,7 @@
 "use client";
 
 import { SessionProvider } from "next-auth/react";
+import React from "react";
 
 type Props = {
   children?: React.ReactNode;
