--- bone-client/utils/db.js
@@ -1,17 +1,0 @@
-import mongoose from "mongoose";
-
-const connect = async () => {
-  if (mongoose.connections[0].readyState) return;
-
-  try {
-    await mongoose.connect(process.env.MONGO_URL, {
-      useNewUrlParser: true,
-      useUnifiedTopology: true,
-    });
-    console.log("Mongo Connection successfully established");
-  } catch (error) {
-    throw new Error("Error connecting to Mongoose");
-  }
-};
-
-export default connect;
